Rota para alteração de senha do usuário cadastrado

+++ routes/manager_user/manager_user_api.py
@@ -7,7 +7,7 @@ from models import User
 from models.user import UserDB
 from db import User as UserTable
 from decorators.is_agent import is_agent
-from schemas.manage_users.manager_users_schema import ManagerUserDelSchema, ManagerUserPostSchema, ManagerUserPutSchema, ManagerUserSchema
+from schemas.manage_users.manager_users_schema import ManagerUserDelSchema, ManagerUserPostSchema, ManagerUserPutSchema, ManagerUserResetPasswdSchema, ManagerUserSchema
 
 def user_to_dict(user: UserTable):
     return {
@@ -65,10 +65,10 @@ class ManagerUserView(MethodView):
     def put(self, item_data):
         '''Atualiza o usuario no sistema'''
 
-        id, name, password, email, agent, active, id_sector = itemgetter('id', 'name', 'password', 'email', 'is_agent', 'active', 'id_sector')(item_data)
+        id, name, email, agent, active, id_sector = itemgetter('id', 'name', 'email', 'is_agent', 'active', 'id_sector')(item_data)
 
         try:
-            update_user: User = UserDB().update_user(id, name, email, password, agent, active, id_sector)
+            update_user: User = UserDB().update_user(id, name, email, agent, active, id_sector)
         except IntegrityError as err:
             abort(400, message='Já existe um usuário registrado com este email')
         except Exception as err:
@@ -91,6 +91,9 @@ class ManagerUserView(MethodView):
 
         id = itemgetter('id')(item_data)
 
+        if id == 1:
+            abort(400, message='O usuário admin não pode ser excluido, caso deseje desative-o')
+
         try:
             UserDB().delete_user(id)
         except Exception as err:
@@ -98,4 +101,26 @@ class ManagerUserView(MethodView):
 
         return {
             'sucesso': 'Usuario excluido com sucesso',
+        }
+
+@blp.route('/manager_user_passwd')
+class ManagerUserPasswdView(MethodView):
+
+    @login_required
+    @is_agent
+    @blp.arguments(ManagerUserResetPasswdSchema)
+    @blp.response(200, ManagerUserResetPasswdSchema)
+    def post(self, item_data):
+        '''Atualiza a senha do usuário no sistema.'''
+
+        id, password = itemgetter('id', 'password')(item_data)
+
+        try:
+            new_user: User = UserDB().set_password(id, password)
+        except Exception as err:
+            abort(400, message='Erro ao tentar atualizar a senha do usuário.')
+
+        return {
+            'data': new_user,
+            'sucesso': 'Senha alterada com sucesso',
         }
